requirements.txt
selenium
beautifulsoup4
requests
webdriver-manager
python-dotenv